requirements.txt
chardet==5.2.0
et-xmlfile==1.1.0
lxml==5.2.1
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
pillow==10.3.0
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
pytz==2024.1
reportlab==4.1.0
six==1.16.0
typing_extensions==4.11.0
tzdata==2024.1
